+++ interludes/secret_example.py
@@ -2,6 +2,8 @@
 
 SECRET_KEY = "i*4$=*fa(644(*!9m2)0-*&sows2uz$b^brb(=)elfn3+y6#1n"
 
+# List of ("name", "email") people who get server notifications:
+# errors, new activities submitted, mass email sent
 ADMINS = [
     ("superuser", "superuser@admin.fr"),
 ]